Section_06/18_Sending Email to multiple persons/pyproject.toml
[project]
name = "18-sending-email-to-multiple-persons"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
]
